--- feature/create_gathering/src/main/java/com/example/create_gathering/navigation/CreateGatheringNavigation.kt
@@ -1,21 +1,0 @@
-package com.example.create_gathering.navigation
-
-import androidx.navigation.NavController
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.compose.composable
-import com.example.create_gathering.CreateGatheringScreen
-import com.plottwist.core.ui.navigation.Route
-
-fun NavGraphBuilder.createGatheringNavGraph(
-    onSubmit: () -> Unit
-) {
-    composable<Route.CreateGathering> {
-        CreateGatheringScreen(
-            onSubmit = onSubmit
-        )
-    }
-}
-
-fun NavController.navigateToCreateGathering() {
-    this.navigate(Route.CreateGathering)
-}
