+++ feature/invite_gathering/src/main/java/com/plottwist/invite_gathering/navigation/InviteGatheringNavigation.kt
@@ -1,0 +1,21 @@
+package com.plottwist.invite_gathering.navigation
+
+import androidx.navigation.NavController
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.compose.composable
+import com.plottwist.invite_gathering.InviteGatheringScreen
+import com.plottwist.core.ui.navigation.Route
+
+fun NavGraphBuilder.inviteGatheringNavGraph(
+    onCloseClicked: () -> Unit
+) {
+    composable<Route.InviteGathering>{
+        InviteGatheringScreen(
+            onCloseClicked = onCloseClicked
+        )
+    }
+}
+
+fun NavController.navigateToInviteGathering() {
+    this.navigate(Route.InviteGathering)
+}
